+++ shuup/notify/signals.py
@@ -8,4 +8,5 @@
 
 from django.dispatch import Signal
 
+notification_email_before_send = Signal(providing_args=["action", "message", "context"])
 notification_email_sent = Signal(providing_args=["message", "context"])