+++ shuup/admin/modules/media/utils.py
@@ -21,6 +21,7 @@ def delete_folder(folder):
     :rtype: str
     """
     parent_folder = (folder.parent if folder.parent_id else None)
+
     parent_name = (parent_folder.name if parent_folder else _("Root"))
     subfolders = list(folder.children.all())
     message_bits = []