+++ CHANGELOG.md
@@ -31,6 +31,11 @@ List all changes after the last release here (newer on top). Each change on a se
 
 ### Added
 
+- Admin: user and permission based access to media folders
+  
+  This means that all vendors can have their own root folder and do what every they want in that folder.
+  But it also allows the admin to give viewing access to one folder for all suppliers.
+
 - Core: add parameter in shuup_static to load the version of a given package