+++ CHANGELOG.md
@@ -8,6 +8,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 List all changes after the last release here (newer on top). Each change on a separate bullet point line.
 
+
+## [2.0.6] - 2020-08-18
+
 ### Changed
 
 - Admin: Make the order editor keep the suppliers of non-product order lines intact.